+++ test/mitmproxy/builtins/test_filestreamer.py
@@ -7,7 +7,7 @@ import os.path
 from mitmproxy.builtins import filestreamer
 from mitmproxy.flow import master, FlowReader
 from mitmproxy.flow import state
-from mitmproxy.flow import options
+from mitmproxy import options
 
 
 class TestStream(mastertest.MasterTest):
